% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airport_data}
\alias{airport_data}
\title{dataframe that contains information about airport codes, latitude and longitude coordinate}
\format{
a dataframe with 354 rows and 3 columns
\describe{
    \item{airport_code}{codes for the airport}
    \item{longitude.degree}{latitude of airport, in decimal degrees north}
    \item{latitude.degree}{longitude of airport, in decimal degrees east}
}
}
\usage{
airport_data
}
\description{
A dataset with the coordinates of 354 airports in the USA
}
\keyword{datasets}
