% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airports_nearby.R
\name{airports_nearby}
\alias{airports_nearby}
\title{find all airports in `airport_data` within `x` miles of a longitude latitude point}
\usage{
airports_nearby(lonlat, airport_data, x = 75)
}
\arguments{
\item{lonlat}{a length-2 vector or a 1 x 2 matrix containing a longitude and latitude}

\item{airport_data}{the airport_data frame object}

\item{x}{within x miles of lonlat, default x=75 miles}
}
\value{
return a data frame that includes the information about airports(airport_code, longitude,
latitude) that are within x miles
}
\description{
given a longitude/latitude coordinate and return the information about airports(airport_code, longitude,
latitude) that are within x miles
}
\examples{
# airports within 75 miles of Ithaca
airports_nearby(matrix(c(-76.45872,42.49136),1,2,byrow=TRUE),airport_data)
# airports within 75 miles of Ithaca, lonlat as length 2 vector
airports_nearby(c(-76.45872,42.49136),airport_data,75)
# airports within 100 miles of Ithaca
airports_nearby(matrix(c(-76.45872,42.49136),1,2,byrow=TRUE),airport_data,100)
}
