% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_symmetric.R
\name{check_symmetric}
\alias{check_symmetric}
\title{test whether a `direct_from` object is symmetric}
\usage{
check_symmetric(df = direct_from)
}
\arguments{
\item{direct_from}{create a list containing information about direct connections from each airport}
}
\value{
return a list whose first element is TRUE or FALSE (whether or not it is symmetric),
and whose second element is an `n x 2` character matrix giving the `n` asymmetricconnections.
}
\description{
test whether a `direct_from` object is symmetric
}
\examples{
# check if list direct_from is symmetric
check_symmetric(direct_from)
}
