% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_matrix.R
\docType{data}
\name{connection_matrix}
\alias{connection_matrix}
\title{A matrix keeps all the information of flights which connect}
\format{
a matrix
}
\usage{
connection_matrix
}
\description{
A matrix keeps all the information of flights which connect
}
\keyword{datasets}
