% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continent_df.R
\docType{data}
\name{continent_df}
\alias{continent_df}
\title{dataframe that contains information about airport codes, latitude and longitude coordinate}
\format{
a dataframe with 309 rows and 3 columns
\describe{
    \item{airport_code}{codes for the airport}
    \item{longitude.degree}{latitude of airport, in decimal degrees north}
    \item{latitude.degree}{longitude of airport, in decimal degrees east}
}
}
\usage{
continent_df
}
\description{
A dataset with the coordinates of 309 airports within lonlat_USA (continental USA)
}
\keyword{datasets}
