% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_from.R
\docType{data}
\name{direct_from}
\alias{direct_from}
\title{a list containing information about direct connections from each airport.}
\format{
a list with length of 354
}
\usage{
direct_from
}
\description{
A list of character containing the airport code from each airport. The list is symmetry, meaning that
if you can fly from airport A to airport B, you are also able to fly from airport B to airport A
}
\keyword{datasets}
