% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_between_airports.R
\docType{data}
\name{dist_between_airports}
\alias{dist_between_airports}
\title{A matrix containing the distance between every pair of airports}
\format{
a matrix of 354 x 354
}
\usage{
dist_between_airports
}
\description{
This matrix is built from airline_2019-07-01.csv which contains all the flights information.
The row names are the names of the airports where flights depart. The column names are the
names of the destination airports. The matrix is not symmetric, which means that A flies to
B doesn't imply that B also flies to A.
}
\keyword{datasets}
