% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great_circle_dist.R
\name{distfun}
\alias{distfun}
\title{compute the great circle distance between pairs of longitude/latitude coordinates}
\usage{
distfun(lonlat1, lonlat2)
}
\arguments{
\item{lonlat1}{an n1 x 2 matrix, first column longitudes, second column latitudes}

\item{lonlat2}{an n2 x 2 matrix, first column longitudes, second column latitudes}
}
\value{
return an n1 x n2 matrix of distances. distmat[i,j] is the distance between
lonlat1[i,] and lonlat2[j,]
}
\description{
given a pair of longtiude/latitude coordinate and return the great circle distance
between two points
}
\examples{
# lonlat1 contains 2 x 2 matrix:
lonlat1 <- matrix(c(-76.10631,43.11119,-87.90815,41.97694),2,2,byrow=TRUE)
# lonlat2 contains 1 x 2 matrix
lonlat2 <- matrix(c(-76.45872,42.49136),1,2,byrow=TRUE)
distfun(lonlat1,lonlat2)
# returns dismat with 2 x 1 matrix
}
