% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat_usa.R
\docType{data}
\name{lonlat_usa}
\alias{lonlat_usa}
\title{A dataframe that contains information about the coordinates of grids in continental
USA.}
\format{
a dataframe with 9285 rows and 2 columns
\describe{
    \item{val1}{longitude of a grid, in decimal degrees north}
    \item{val2}{latitude of a grid, in decimal degrees east}
}
}
\usage{
lonlat_usa
}
\description{
A dataframe that contains information about the coordinates of grids in continental
USA.
}
\keyword{datasets}
