% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachable_points.R
\name{reachable_points}
\alias{reachable_points}
\title{Compute the points reachable from a lonlat coordinate by traveling less than 2x miles}
\usage{
reachable_points(data, lonlat, x, df = direct_from)
}
\arguments{
\item{data}{a n1 x 2 dataframe, first column longitude, second column latitudes}

\item{lonlat}{an n2 x 2 matrix, first column longitudes, second column latitudes}

\item{x}{a positive number}

\item{df}{a list containing the direct_from information}
}
\value{
return an n3 x 2 dataframe containing all the grids reachable from the given
coordinate
}
\description{
Given a pair of longtiude/latitude coordinate and return the grids reachable from
this coordinate.
}
